Extrais le contenu de l'onglet mesure dans un composant
